+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 Download at [Maven Search](https://search.maven.org/search?q=g:de.lars-sh%20AND%20%28a:parent%20a:parent-archetype%20a:root%20a:utils%20a:utils-annotations%20a:utils-test%29%20AND%20v:0.9.5)
 
 ### Added
+* Check code format via command line prior compiling. This simplifies working in teams with different editors.
 * Method de.larssh.utils.Either.ifPresent(firstConsumer, secondConsumer)
 * Method de.larssh.utils.Either.ifFirstIsPresent(consumer)
 * Method de.larssh.utils.Either.ifSecondIsPresent(consumer)