+++ CHANGELOG.md
@@ -17,8 +17,9 @@ Download at [Maven Search](https://search.maven.org/search?q=g:de.lars-sh%20AND%
 * Checkstyle nearby comment filter for the MultipleStringLiterals check. Use `@SuppressWarnings` instead.
 
 ### Changed
-* Avoid some warnings inside the Maven output.
+* Moved execution of the `tidy-maven-plugin` to `validate` and creation of project files to the Maven lifecycle phase `initialize`.
 * Checkstyle no longer runs on delomboked sources.
+* Avoid some ignorable warnings inside the Maven output.
 
 ### Fixed
 * TODO