+++ CHANGELOG.md
@@ -13,10 +13,18 @@ Download at [Maven Search](https://search.maven.org/search?q=g:de.lars-sh%20AND%
 * Method de.larssh.utils.Either.ifPresent(firstConsumer, secondConsumer)
 * Method de.larssh.utils.Either.ifFirstIsPresent(consumer)
 * Method de.larssh.utils.Either.ifSecondIsPresent(consumer)
-* POM properties to configure generation of files and folders
+* Additional POM properties to configure generation of files and folders
 
 ### Removed
 * Checkstyle nearby comment filter for the MultipleStringLiterals check. Use `@SuppressWarnings` instead.
+* Renamed the Maven property `org.eclipse.jdt.core.compiler.annotation.missingNonNullByDefaultAnnotation`. Use `eclipse.compiler.missingNonNullByDefaultAnnotation` instead.
+* Renamed the Maven property `org.eclipse.jdt.core.compiler.annotation.nonnull.secondary`. Use `eclipse.compiler.nonnull.secondary` instead.
+* Renamed the Maven property `org.eclipse.jdt.core.compiler.annotation.nonnullbydefault.secondary`. Use `eclipse.compiler.nonnullbydefault.secondary` instead.
+* Renamed the Maven property `org.eclipse.jdt.core.compiler.annotation.nullable.secondary`. Use `eclipse.compiler.nullable.secondary` instead.
+* Renamed the Maven property `org.eclipse.jdt.core.compiler.doc.comment.support`. Use `eclipse.compiler.javadoc` instead.
+* Renamed the Maven property `project.build.mainClass`. Use `jar.manifest.mainClass` instead.
+* Renamed the Maven property `project.build.packageDependenciesPhase`. Use `shade.packageDependenciesPhase` instead.
+
 
 ### Changed
 * Moved execution of the `tidy-maven-plugin` to `validate` and creation of project files to the Maven lifecycle phase `initialize`.