+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 <a name="0.9.5"></a>
 
-## [0.9.5](https://github.com/lars-sh/parent/compare/35490fbfcecc4c4eace0d3d3f630bd121bd0878a...master) (2019-12-03)
+## [0.9.5](https://github.com/lars-sh/parent/compare/35490fbfcecc4c4eace0d3d3f630bd121bd0878a...910893179255a2a139856edf72bc5525bb9f3e40) (2019-12-03)
 
 Download at [Maven Search](https://search.maven.org/search?q=g:de.lars-sh%20AND%20%28a:parent%20a:parent-archetype%20a:root%20a:utils%20a:utils-annotations%20a:utils-test%29%20AND%20v:0.9.5)